requirements.txt
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
rich==13.8.1
shellingham==1.5.4
speedtest-cli==2.1.3
typer==0.12.5
typing_extensions==4.12.2
